fix: use float inputs instead of double

+++ src/ros/rover/arm/arm_kinematics/arm_kinematics/include/arm_kinematics/collision/collision_manager.hpp
@@ -24,7 +24,7 @@ struct ARM_KINEMATICS_PUBLIC CollisionManager {
     DiscreteCollisionPlugin::SharedPtr plugin);
 
   [[nodiscard]] bool collide() const;
-  void update_poses(const std::vector<double> & joint_states);
+  void update_poses(const std::vector<float> & joint_states);
 
 private:
   ForwardKinematicsPlugin::Tree::SharedPtr tree_{};
